api: Remove unnecessary action method prefix
None of the other controllers do this. Don't do it here either. This is
mostly a bulk rename, with the exception of the combination of the
'_action_resize' and '_resize' methods, which were unnecessarily
separated, and the move of the 'delete' method to be next to the
'_delete' inner method.

Change-Id: I87381c6721e7a040c82f8124523116a1d4e2c684
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>